+++ example/main.tf
@@ -5,21 +5,22 @@
 module "account_example-account_dev" {
   source    = "../"
 
+  org_name                = "my-org"
   region                  = "us-east-1"
   master_account_id       = "01234567891"
   account_name            = "account-1"
   email_owner             = "example-account_dev@example.com"
   organizational_unit_id  = "ou-1234-12345678"
   role_name               = "OrgDeploy"
-  ebs_default_kms_key_arn = "arn:aws:kms:eu-west-2:123456789012:key/123"
+
+  block_s3_public_access        = true
+  enable_default_ebs_encryption = true
 
   budget_limit_amount = "1000.0"
   budget_emails = [
     "user1@example.com",
     "user2@example.com",
   ]
-
-  iam_namespace = "my-org"
 
   permitted_services = [
     "dynamodb",